feat: introduced an extended version of the FileServlet
